requirements.txt
# Web 框架和服务器
fastapi>=0.104.0
uvicorn>=0.24.0
starlette>=0.27.0

# HTTP 客户端
requests>=2.31.0
httpx>=0.25.0

# OpenAI 客户端
openai>=1.3.0

# 环境变量管理
python-dotenv>=1.0.0

# 类型提示支持
typing-extensions>=4.8.0

# 测试工具
pytest>=7.4.3

# JSON 处理
pydantic>=2.4.2

# 异步支持
asyncio>=3.4.3

# 开发工具（可选）
black>=23.10.0  # 代码格式化
flake8>=6.1.0   # 代码检查
isort>=5.12.0   # import 排序

# 日志管理（可选）
loguru>=0.7.2